refactor: 랜덤 생성 로또 번호 분리
