+++ src/git.txt
@@ -4,7 +4,7 @@ git remote add origin https://github.com/jngbnss/preCoursePractice.git
 git status
 
 git add .
-git commit -m "refactor: 로또 번호 및 보너스 번호 저장"
+git commit -m "refactor: 랜덤 생성 로또 번호 분리"
 git push origin jngbnss
 
 git pull origin main
